Test aggressive optimization with GCC 7.5.0.
If you replace '-g' with '-O3' in the original 'Makefile' of the
original 'sgb.tar.gz' tarball, './test_io' fails with this error:

    Bad checksum, or difficulty closing the file!
    (Error code = 512)

Debugging shows that it suffices to 'gcc -O3 -c gb_io.c' to 'make tests'
fail.

Optimization with '-O' and '-O2' works fine.

Curiously enough, the problem with '-O3' does _not_ occur when we build
the dynamic version of 'sgb.so' instead of the static version 'sgb.a'.

Hence this change in 'Makefile', which hinges on the extreme edge of
possiible failure.